+++ playground/ab/parser/parser.h
@@ -3,6 +3,9 @@
 
 # include <stdio.h>
 
+typedef struct s_deque_node	t_deque_node;
+typedef struct s_deque		t_deque;
+
 typedef enum e_node_kind {
 	NODE_ADD = 0,
 	NODE_SUB = 1,
@@ -17,5 +20,9 @@ typedef struct s_node {
 	struct s_node	*right;
 	int				val; // use only kind == NODE_NUM
 }	t_node;
+
+// token_new.c
+t_node	*node_new(t_node_kind kind, t_node *left, t_node *right);
+t_node	*node_num_new(int val);
 
 #endif
